+++ index.js
@@ -3,7 +3,7 @@
 /**
  * Router
  *
- * @version: 1.1.0
+ * @version: 1.1.2
  * @author Graidenix
  *
  * @constructor
@@ -176,6 +176,7 @@ Router.prototype._parseRouteRule = function (route) {
  * @private
  */
 Router.prototype._parseQuery = function (query) {
+    this._queryString = query;
     var _query = {};
     if (typeof query !== 'string') {
         return _query;
@@ -324,11 +325,6 @@ Router.prototype.check = function () {
     var self = this,
         fragment = this._getFragment();
 
-    if (this._skipCheck) {
-        this._skipCheck = false;
-        return;
-    }
-
     if (!this._unloadCallback()) {
         this._skipCheck = true;
         if (this.mode === 'history') {
@@ -349,8 +345,12 @@ Router.prototype.check = function () {
             var query = self._getQuery();
             var page = new Router.Page(fragment, query, match, self._pageState, route.options);
 
-            self.beforeHook(page);
             self._currentPage = page;
+            if (self._skipCheck) {
+                self._skipCheck = false;
+                return true;
+            }
+            self.beforeHook(page);
             route.handler.apply(page, match);
             self._pageState = null;
 
@@ -409,12 +409,14 @@ Router.prototype.removeUriListener = function () {
  *
  * @param {string} path
  * @param {object} state
+ * @param {boolean} silent
  *
  * @returns {Router}
  */
-Router.prototype.navigateTo = function (path, state) {
+Router.prototype.navigateTo = function (path, state, silent) {
     path = this._trimSlashes(path) || '';
     this._pageState = state || null;
+    this._skipCheck = !!silent;
     if (this.mode === 'history') {
         history.pushState(state, null, this.root + this._trimSlashes(path));
         return this.check();
@@ -422,6 +424,15 @@ Router.prototype.navigateTo = function (path, state) {
         window.location.hash = path;
     }
     return this;
+};
+
+/**
+ * Refresh page with recall route handler
+ * @returns {Router}
+ */
+Router.prototype.refresh = function () {
+    var path = this._currentPage.uri + '?' + this._queryString;
+    return this.navigateTo(path, this._currentPage.state);
 };
 
 /**